fix: use 16-bit IDs to avoid LR node ID truncation
Transport Service ZIPGW bridge path was narrowing the node IDs from
16-bit to 8-bits.

Add a standalong test that uses the flags used to build the Z/IP Gateway
and avoid the specific NEW_TEST_T2, as the test would result in
different `ZW_SendData_Bridge` object.

Use CMock to simplify mocking and to easy eventual new tests.

Relates-to: ZGW-3457
Origin: https://github.com/SiliconLabs/zipgateway/pull/47

+++ libs2/transport_service/transport_service2.c
@@ -107,7 +107,7 @@ typedef void (*ZW_CommandHandler_Callback_t)(ts_param_t* p, ZW_APPLICATION_TX_BU
 #endif
 
 #ifndef __C51__
-extern uint8_t ZW_SendData_Bridge(uint8_t, uint8_t, uint8_t *, uint8_t, uint8_t, VOID_CALLBACKFUNC(completedFunc)(uint8_t, TX_STATUS_TYPE*));
+extern uint8_t ZW_SendData_Bridge(uint16_t, uint16_t, uint8_t *, uint8_t, uint8_t, VOID_CALLBACKFUNC(completedFunc)(uint8_t, TX_STATUS_TYPE*));
 #endif /* ifndef __C51__ */
 
 #endif
